rag_chatbot/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
faiss-cpu
google-generativeai
pypdf
tqdm
numpy
markdown
python-multipart
